A. task_cfg.py: 1. (822, 866, 1442, 1074-1087) extract function get_rsc_2b_released by the way fix bug, all the rsc should be released after the point of context switching, rather than the section of conflict happens; 2. rename issue_process as allocate_rsc_4_process; 3. (1144-1268, 1368-1495) extract function preempt_the_conflicts: 1393-1405 additional evaluation of the preemption possibility (ensure the preemptable conflict tasks can provide enough FLOPS to the current task); add key to filter the confict tasks that have lower affinity with the current bin compared with the current task

+++ scheduling_table.py
@@ -138,7 +138,7 @@ class SchedulingTableInt(object):
             curr_slot = np.zeros(len(s), dtype=int)
 
             # avalable (A)
-            rsc_avl_tmp = np.zeros(len(s)) 
+            rsc_avl_tmp = np.zeros(len(s), dtype=int) 
 
             # required (R)
             expected_req_rsc_size = req_rsc_size * expected_slot_num
